add facebook social network, moved header
