+++ src/assets/css/2020/common.scss
@@ -1,10 +1,5 @@
 @import './var.scss';
 
-
-body{
-  background-color: $background-color !important;
-}
-
 .section{
 
   &-bg-white{
